EA-97 | SortWeight For Orders In OrderGroups:
Exposed sortWeight field, so that clients can set the sortWeights

+++ api/src/main/java/org/openmrs/module/emrapi/encounter/domain/EncounterTransaction.java
@@ -874,7 +874,7 @@ public class EncounterTransaction {
         private String voidReason;
         private EncounterTransaction.Concept orderReasonConcept;
 
-        private Integer sortWeight;
+        private Double sortWeight;
 
         public Drug getDrug() {
             return drug;
@@ -995,11 +995,11 @@ public class EncounterTransaction {
             return voidReason;
         }
 
-        public Integer getSortWeight() {
+        public Double getSortWeight() {
             return sortWeight;
         }
 
-        public void setSortWeight(Integer sortWeight) {
+        public void setSortWeight(Double sortWeight) {
             this.sortWeight = sortWeight;
         }
     }
